# swagger.yaml
openapi: 3.0.3
info:
  title: Menu Link API
  description: |
    API backend pour l'application SaaS de gestion de restaurant avec menu digital.
    
    ## Fonctionnalités
    - Multi-Restaurants (SaaS)
    - Gestion du compte Gérant et du Staff (Caissiers, Cuisine, Serveurs)
    - Prise de commande par QR code 
    - Séparation des vues (Client public vs Staff privé)
    
    ## Authentification
    Cette API utilise un système de jetons JWT. Un bouton "Authorize" se trouve ci-dessous pour inclure votre JWT dans les requêtes protégées. Les endpoints publics (ex: vue du menu depuis un QR code) ne nécessitent pas de token.
    
  version: 1.0.0
servers:
  - url: https://menulink-api.duckdns.org/api
    description: Production server (AWS EC2)
  - url: https://menu-link-api.onrender.com/api
    description: Backup server (Render)
  - url: http://localhost:3000/api
    description: Development server
...
paths:
...
  /sessions/{session_id}:
    get:
      tags:
        - Sessions
      summary: Validate session
      description: Vérifie si une session est toujours valide
      operationId: validateSession
      parameters:
        - name: session_id
          in: path
          required: true
          description: ID de la session à valider
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '200':
          description: Session valide
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      session_id:
                        type: string
                        format: uuid
                        example: "550e8400-e29b-41d4-a716-446655440000"
                      table_id:
                        type: string
                        example: "507f1f77bcf86cd799439011"
                      table_number:
                        type: integer
                        example: 1
                      started_at:
                        type: string
                        format: date-time
                        example: "2024-01-26T10:30:00.000Z"
                  message:
                    type: string
                    example: "Session is valid"
        '401':
          description: Session invalide ou expirée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                success: false
                error:
                  code: "INVALID_SESSION"
                  message: "Invalid or expired session"
    
    delete:
      tags:
        - Sessions
      summary: End session
      description: Termine une session client
      operationId: endSession
      parameters:
        - name: session_id
          in: path
          required: true
          description: ID de la session à terminer
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '200':
          description: Session terminée avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Session ended successfully"
        '404':
          description: Session non trouvée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /menus/{id}:
    get:
      tags:
        - Menu
      summary: Get menu by ID
      description: Récupère un menu spécifique par son ID
      operationId: getMenuById
      parameters:
        - name: id
          in: path
          required: true
          description: ID du menu
          schema:
            type: string
            example: "507f1f77bcf86cd799439011"
      responses:
        '200':
          description: Menu trouvé
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Menu'
        '404':
          description: Menu non trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    put:
      tags:
        - Menu
      summary: Update menu
      description: Met à jour un menu existant
      operationId: updateMenu
      parameters:
        - name: id
          in: path
          required: true
          description: ID du menu
          schema:
            type: string
            example: "507f1f77bcf86cd799439011"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "Menu Spécial"
                is_active:
                  type: boolean
                  example: false
      responses:
        '200':
          description: Menu mis à jour
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Menu'
                  message:
                    type: string
                    example: "Menu updated successfully"
    
    delete:
      tags:
        - Menu
      summary: Delete menu
      description: Supprime un menu
      operationId: deleteMenu
      parameters:
        - name: id
          in: path
          required: true
          description: ID du menu
          schema:
            type: string
            example: "507f1f77bcf86cd799439011"
      responses:
        '200':
          description: Menu supprimé
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Menu deleted successfully"
  /menus:
    post:
      tags:
        - Menu
      summary: Create menu
      description: Crée un nouveau menu
      operationId: createMenu
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - manager_id
              properties:
                name:
                  type: string
                  example: "Menu du Jour"
                manager_id:
                  type: string
                  example: "507f1f77bcf86cd799439010"
      responses:
        '201':
          description: Menu créé
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Menu'
                  message:
                    type: string
                    example: "Menu created successfully"
        '400':
          description: Champs manquants
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: "ERROR_CODE"
            message:
              type: string
              example: "Error description"
            details:
              type: object
              additionalProperties: true
    Menu:
      type: object
      properties:
        _id:
          type: string
          example: "507f1f77bcf86cd799439011"
        name:
          type: string
          example: "Menu Principal"
        is_active:
          type: boolean
          example: true
        manager_id:
          type: string
          example: "507f1f77bcf86cd799439010"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []